refactor: rename .entrypoint to .installer
