+++ __tests__/unit/categories/js/eslint/eslint.installer.spec.ts
@@ -1,9 +1,9 @@
 import { presetState, fileSystem, terminal } from '#__tests__/setup-test-context.ts';
 import { createConfig } from '#__tests__/utils/eslint.utils.ts';
-import { buildConfig, eslint } from '#src/categories/js/eslint/eslint.entrypoint.ts';
+import { buildConfig, eslint } from '#src/categories/js/eslint/eslint.installer.ts';
 import { beforeEach, describe, expect, it } from 'vitest';
 
-describe('eslint.entrypoint', () => {
+describe('eslint.installer', () => {
   beforeEach(() => {
     presetState.setJsPreset('default');
   });
