+++ src/categories/js/dependabot/dependabot.installer.ts
