+++ src/categories/js/docker/docker.installer.ts
