+++ src/categories/js/eslint/eslint.installer.ts
