+++ src/categories/js/husky/husky.installer.ts
