+++ src/categories/js/jest/jest.installer.ts
