+++ src/categories/js/lint-staged/lint-staged.installer.ts
