+++ src/categories/js/stylelint/stylelint.installer.ts
